FTPServer.java : Addition of the path of the repository that will be manipulate by the server.
To do it I had to import the java.io.File but there was already the import of the class util.File. That caused so troubles with the name so I removed the import of the util.File class et specify directly into the code where it's need to call method from it.