FTPRequest.java : Bug correction
While the connection with the server once with send a valide username, we could be authentified no matter the password we said.

The problem come from the method which check the identity. It compared the given username with the password stored in the csvfile. However we stored only two users with their login as password thanks to that these two login didn't need to know their password to log in.